chore: experimenting with async aborter
For now, I'm going to stop pursuing this approach in favor of using Promise.race(). There's not much support for it outside of Node.js internals unless you run code in a separate thread or process, which wouldn't help with sharing a vm.Context throughout all statement executions. Definitely something to explore more in the future.